streamlit-app/requirements.txt
streamlit==1.45.1
pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
